refactor: moved projects and experience to sub folders
