+++ src/Images/Projects/algorithm-animation.png
